Split GraphQL scalar code generation
