+++ webserver/graphql/src/main/java/io/helidon/webserver/graphql/GraphQlService.java
@@ -40,8 +40,8 @@ import io.helidon.common.types.TypedElementInfo;
 import io.helidon.common.uri.UriQuery;
 import io.helidon.config.Config;
 import io.helidon.graphql.server.ExecutionContext;
-import io.helidon.graphql.server.GraphQlContextKeys;
 import io.helidon.graphql.server.GraphQlConstants;
+import io.helidon.graphql.server.GraphQlContextKeys;
 import io.helidon.graphql.server.InvocationHandler;
 import io.helidon.http.HeaderNames;
 import io.helidon.http.Status;
